nixos/zfs: add keyScript option
Previously, dataset credentials were only able to be provided through a file on disk or a prompted password.
This is problematic for systems which want to make use of tools such as [clevis](https://github.com/latchset/clevis) to bind datasets to TPMs.
Clevis is included in this PR as a reference example.
These changes introduce the concepet of key scripts, which are run before attempting an interactive password prompt and are able to unlock the dataset.